Feature/release 2.6

+++ src/Sfa.Tl.Matching.Web/Controllers/NavigationController.cs
@@ -21,14 +21,17 @@ namespace Sfa.Tl.Matching.Web.Controllers
 
         [HttpGet]
         [Route("remove-opportunityItem/{opportunityId}-{opportunityItemId}", Name = "RemoveAndGetOpportunityBasket")]
-        public async Task<IActionResult> RemoveOpportunityItemAndGetOpportunityBasketAsync(int opportunityId, int opportunityItemId)
+        public async Task<IActionResult> RemoveAndGetOpportunityBasketAsync(int opportunityId, int opportunityItemId)
         {
-            await _opportunityService.DeleteOpportunityItemAsync(opportunityId, opportunityItemId);
             var opportunityItemCount = await _opportunityService.GetSavedOpportunityItemCountAsync(opportunityId);
 
-            return opportunityItemCount == 0
-                ? RedirectToRoute("Start")
-                : RedirectToRoute("GetOpportunityBasket", new { opportunityId, opportunityItemId });
+            if (opportunityItemCount == 0)
+            {
+                await _opportunityService.DeleteOpportunityItemAsync(opportunityId, opportunityItemId);
+                return RedirectToRoute("Start");
+            }
+
+            return RedirectToRoute("GetOpportunityBasket", new { opportunityId, opportunityItemId });
         }
 
         [HttpGet]
@@ -81,7 +84,7 @@ namespace Sfa.Tl.Matching.Web.Controllers
 
         [HttpGet]
         [Route("get-back-link/{OpportunityId}/{OpportunityItemId}/{Postcode}/{SelectedRouteId}", Name = "GetBackLink")]
-        public async Task<IActionResult> BackLink(SearchParametersViewModel viewModel)
+        public async Task<IActionResult> GetBackLinkAsync(SearchParametersViewModel viewModel)
         {
             var prevUrl = await _backLinkService.GetBackLinkAsync(HttpContext.User.GetUserName());
 
@@ -101,5 +104,21 @@ namespace Sfa.Tl.Matching.Web.Controllers
 
             return Redirect(prevUrl);
         }
+
+        [HttpGet]
+        [Route("get-admin-back-link/{ProviderId}", Name = "GetAdminBackLink")]
+        public async Task<IActionResult> GetAdminBackLinkAsync(BackLinkViewModel viewModel)
+        {
+            var prevUrl = await _backLinkService.GetBackLinkAsync(HttpContext.User.GetUserName());
+
+            if (prevUrl.Contains("create-provider") && viewModel.ProviderId != 0)
+            {
+                await _backLinkService.GetBackLinkAsync(HttpContext.User.GetUserName());
+                return RedirectToRoute("GetProviderDetail", viewModel.ProviderId);
+            }
+
+            return Redirect(prevUrl);
+        }
+
     }
 }
